requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0,<2
scipy>=1.14.1,<1.15  # 1.15+/1.16+ can raise "All ufuncs must have type numpy.ufunc" with numpy 1.26
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
Pillow>=9.0.0
PyYAML>=6.0
tqdm>=4.62.0

# Optional but recommended
onnx>=1.12.0
onnxruntime>=1.12.0
timm>=0.6.0  # For EfficientNet if torchvision version is old

# Optional for FLOPs estimation
# thop>=0.1.0
# ptflops>=0.1.0

# Optional for robustness evaluation (motion blur)
# opencv-python>=4.5.0

